requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
flask>=3.0.2
Tensorflow>=2.15.0
numpy>=1.26.4
opencv-python>=1.26.4
pillow>=10.2.0
